Initial implementation removing Lapin.Worker
Remove Lapin.Worker in favour of Lapin.Connection as a behaviour for
worker logic. Decouple Lapin.Pattern and configure it per channel.

Also generate a publish shorthand in the worker module to address #2
