+++ lib/lapin/connection.ex
@@ -3,8 +3,9 @@ defmodule Lapin.Connection do
   RabbitMQ connection handler
   """
   use AMQP
-  use GenServer
   require Logger
+
+  use GenServer
 
   alias Lapin.{Message, Worker}
 
@@ -34,14 +35,10 @@ defmodule Lapin.Connection do
   @type routing_key :: String.t
 
   @typedoc """
-  Channel role
-  """
-  @type role :: :consumer | :producer | :passive
-
-  @typedoc """
   Connection configuration
 
   The following keys are supported:
+    - module: module adopting the `Lapin.Connection` behaviour
     - host: broker hostname (string | charlist), *default: 'localhost'*
     - port: broker port (string | integer), *default: 5672*
     - virtual_host: broker vhost (string), *default: ""*
@@ -58,10 +55,10 @@ defmodule Lapin.Connection do
 
   The following keys are supported:
     - role: channel role (`atom`), allowed values are:
-      - `:consumer`: Receives messages from the channel via the `worker` module
+      - `:consumer`: Receives messages from the channel via `Lapin.Connection` callbacks
       - `:producer`: Can publish messages to che channel
       - `:passive`: Used to declare channel configuration, can't receive nor publish
-    - worker: channel worker (module conforming to the `Lapin.Worker` behaviour)
+    - pattern: channel pattern (module adopting the `Lapin.Pattern` behaviour)
     - exchange: broker exchange (`String.t`)
     - queue: broker queue (`String.t`)
 
@@ -79,12 +76,96 @@ defmodule Lapin.Connection do
   """
   @type channel_config :: Keyword.t
 
+  @typedoc """
+  `Lapin.Connection` generic callback result
+  """
+  @type on_callback :: :ok | {:error, message :: String.t}
+
+  @typedoc """
+  Reason for message rejection
+  """
+  @type reason :: term
+
+  @typedoc """
+  `Lapin.Conenction` handle_deliver callback result
+  """
+  @type on_deliver :: :ok | {:reject, reason} | {:requeue, reason} | term
+
+  @doc """
+  Called when receiving a `basic.cancel` from the broker.
+  """
+  @callback handle_cancel(channel_config :: Connection.channel_config) :: on_callback
+
+  @doc """
+  Called when receiving a `basic.cancel_ok` from the broker.
+  """
+  @callback handle_cancel_ok(channel_config :: Connection.channel_config) :: on_callback
+
+  @doc """
+  Called when receiving a `basic.deliver` from the broker.
+
+  Return values from this callback determine message acknowledgement:
+    - `:ok`: Message was processed by the consumer and should be removed from queue
+    - `{:requeue, reason}`: Message was not processed and should be requeued
+    - `{:reject, reason}`: Message was not processed but should NOT be requeued
+
+  Any other return value, including a crash in the callback code, has the same
+  effect as `{:reject, reason}`: it rejects the message WITHOUT requeueing. The
+  `reason` term can be used by the application to signal the reason of rejection
+   and is logged in debug.
+  """
+  @callback handle_deliver(channel_config :: Connection.channel_config, message :: Message.t) :: on_deliver
+
+  @doc """
+  Called when receiving a `basic.consume_ok` from the broker.
+
+  This signals successul registration as a consumer.
+  """
+  @callback handle_consume_ok(channel_config :: Connection.channel_config) :: on_callback
+
+  @doc """
+  Called when completing a `basic.publish` with the broker.
+
+  Message transmission to the broker is successful when this callback is called.
+  """
+  @callback handle_publish(channel_config :: Connection.channel_config, message :: Message.t) :: on_callback
+
+  @doc """
+  Called when receiving a `basic.return` from the broker.
+
+  THis signals an undeliverable returned message from the broker.
+  """
+  @callback handle_return(channel_config :: Connection.channel_config, message :: Message.t) :: on_callback
+
+  defmacro __using__(_) do
+    quote do
+      alias Lapin.Connection, as: Conn
+      alias Lapin.Message
+
+      @behaviour Lapin.Connection
+
+      def handle_cancel(_consumer_tag), do: :ok
+      def handle_cancel_ok(_consumer_tag), do: :ok
+      def handle_consume_ok(_channel_config), do: :ok
+      def handle_deliver(_channel_config, _message), do: :ok
+      def handle_publish(_channel_config, _message), do: :ok
+      def handle_return(_channel_config, _message), do: :ok
+
+      defoverridable Lapin.Connection
+
+      def publish(exchange, routing_key, message, options \\ []) do
+        Conn.publish(__MODULE__, exchange, routing_key, message, options)
+      end
+
+    end
+  end
+
+  @connection_mandatory_params [:module]
+  @channel_mandatory_params [:role, :pattern, :exchange, :queue]
+  @connection_default_params [connecion_timeout: 5_000]
   @default_rabbitmq_host 'localhost'
   @default_rabbitmq_port 5672
   @default_reconnection_delay 5_000
-  @connection_default_params [connecion_timeout: 5_000]
-  @connection_mandatory_params [:handle]
-  @channel_mandatory_params [:role, :worker, :exchange, :queue]
 
   @doc """
   Starts a `Lapin.Connection` with the specified configuration
@@ -115,19 +196,20 @@ defmodule Lapin.Connection do
     GenServer.call(connection, {:publish, exchange, routing_key, message, options})
   end
 
-  def handle_call({:publish, exchange, routing_key, message, options}, _from, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, exchange, routing_key),
+  def handle_call({:publish, exchange, routing_key, message, options}, _from,
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, exchange, routing_key),
          :producer <- Keyword.get(channel_config, :role),
          channel when not is_nil(channel) <- Keyword.get(channel_config, :channel),
-         worker <- Keyword.get(channel_config, :worker),
-         pattern <- worker.pattern(),
+         pattern <- Keyword.get(channel_config, :pattern),
          mandatory <- pattern.publisher_mandatory(channel_config),
          persistent <- pattern.publisher_persistent(channel_config),
          options <- Keyword.merge([mandatory: mandatory, persistent: persistent], options),
          :ok <- Basic.publish(channel, exchange, routing_key, message.payload, options) do
       if not pattern.publisher_confirm(channel_config) or Confirm.wait_for_confirms(channel) do
           Logger.debug fn -> "Published to '#{exchange}'->'#{routing_key}': #{inspect message}" end
-          {:reply, worker.handle_publish(channel_config, message), state}
+          {:reply, module.handle_publish(channel_config, message), state}
         else
           error = "Error publishing #{inspect message} to #{exchange}: broker did not confirm reception"
           Logger.debug fn -> error end
@@ -152,10 +234,11 @@ defmodule Lapin.Connection do
     end
   end
 
-  def handle_info({:basic_cancel, %{consumer_tag: consumer_tag}}, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
-         worker <- Keyword.get(channel_config, :worker),
-         :ok <- worker.handle_cancel(channel_config) do
+  def handle_info({:basic_cancel, %{consumer_tag: consumer_tag}},
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
+         :ok <- module.handle_cancel(consumer_tag) do
         Logger.debug fn -> "Broker cancelled consumer_tag '#{consumer_tag}' for channel #{inspect channel_config}" end
     else
       nil ->
@@ -166,10 +249,11 @@ defmodule Lapin.Connection do
     {:stop, :normal, state}
   end
 
-  def handle_info({:basic_cancel_ok, %{consumer_tag: consumer_tag}}, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
-         worker <- Keyword.get(channel_config, :worker),
-         :ok <- worker.handle_cancel_ok(channel_config) do
+  def handle_info({:basic_cancel_ok, %{consumer_tag: consumer_tag}},
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
+         :ok <- module.handle_cancel_ok(consumer_tag) do
       Logger.debug fn -> "Broker confirmed cancel consumer_tag '#{consumer_tag}' for channel #{inspect channel_config}" end
     else
       nil ->
@@ -180,10 +264,11 @@ defmodule Lapin.Connection do
     {:noreply, state}
   end
 
-  def handle_info({:basic_consume_ok, %{consumer_tag: consumer_tag}}, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
-         worker <- Keyword.get(channel_config, :worker),
-         :ok <- worker.handle_consume_ok(channel_config) do
+  def handle_info({:basic_consume_ok, %{consumer_tag: consumer_tag}},
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
+         :ok <- module.handle_consume_ok(consumer_tag) do
         Logger.debug fn -> "Broker registered consumer_tag '#{consumer_tag}' for channel #{inspect channel_config}" end
     else
       nil ->
@@ -194,12 +279,14 @@ defmodule Lapin.Connection do
     {:noreply, state}
   end
 
-  def handle_info({:basic_deliver, payload, %{consumer_tag: consumer_tag} = meta}, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
+  def handle_info({:basic_deliver, payload, %{consumer_tag: consumer_tag} = meta},
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, consumer_tag),
          channel when not is_nil(channel) <- Keyword.get(channel_config, :channel),
          message <- %Message{meta: meta, payload: payload} do
       spawn(fn ->
-        consume(channel_config, channel, message)
+        consume(module, channel, channel_config, message)
       end)
     else
       nil ->
@@ -208,10 +295,11 @@ defmodule Lapin.Connection do
     {:noreply, state}
   end
 
-  def handle_info({:basic_return, payload, %{exchange: exchange, routing_key: routing_key} = meta}, %{channels: channels} = state) do
-    with channel_config when not is_nil(channel_config) <- get_channel_config(channels, exchange, routing_key),
-         worker <- Keyword.get(channel_config, :worker),
-         :ok <- worker.handle_return(channel_config, %Message{meta: meta, payload: payload}) do
+  def handle_info({:basic_return, payload, %{exchange: exchange, routing_key: routing_key} = meta},
+  %{channels: channels, configuration: configuration} = state) do
+    with module <- Keyword.get(configuration, :module),
+         channel_config when not is_nil(channel_config) <- get_channel_config(channels, exchange, routing_key),
+         :ok <- module.handle_return(channel_config, %Message{meta: meta, payload: payload}) do
       Logger.debug fn -> "Returned message for '#{exchange}'->'#{routing_key}': #{inspect meta}" end
     else
       error ->
@@ -234,12 +322,12 @@ defmodule Lapin.Connection do
     Connection.close(connection)
   end
 
-  defp consume(channel_config, channel, %Message{meta: %{delivery_tag: delivery_tag}} = message) do
+  defp consume(module, channel, channel_config, %Message{meta: %{delivery_tag: delivery_tag}} = message) do
     Logger.debug fn -> "Consuming message #{delivery_tag}" end
-    with worker <- Keyword.get(channel_config, :worker),
-         pattern <- worker.pattern(),
+
+    with pattern = Keyword.get(channel_config, :pattern),
          consumer_ack <- pattern.consumer_ack(channel_config),
-         :ok <- worker.handle_deliver(channel_config, message) do
+         :ok <- module.handle_deliver(channel_config, message) do
       consume_ack(consumer_ack, channel, delivery_tag)
      else
       {:reject, reason} ->
@@ -292,8 +380,7 @@ defmodule Lapin.Connection do
          role when not is_nil(role) <- Keyword.get(channel_config, :role),
          exchange when not is_nil(exchange) <- Keyword.get(channel_config, :exchange),
          queue when not is_nil(queue) <- Keyword.get(channel_config, :queue),
-         worker <- Keyword.get(channel_config, :worker),
-         pattern <- worker.pattern(),
+         pattern <- Keyword.get(channel_config, :pattern, Lapin.Pattern.Config),
          exchange_type <- pattern.exchange_type(channel_config),
          exchange_durable <- pattern.exchange_durable(channel_config),
          queue_arguments <- pattern.queue_arguments(channel_config),
