--- lib/lapin/connection/registry.ex
@@ -1,13 +1,0 @@
-defmodule Lapin.Connection.Registry do
-  @moduledoc """
-  Connection process registry
-  """
-
-  @type handle :: term
-
-  @doc """
-  Returns a via tuple for `Lapin.Connection` process
-  """
-  @spec via(handle) :: {:via, Registry, {__MODULE__, handle}}
-  def via(handle), do: {:via, Registry, {__MODULE__, handle}}
-end
