+++ test/test_helper.exs
@@ -1,9 +1,5 @@
 ExUnit.start()
 
-defmodule LapinTest.HelloWorldWorker do
-  use Lapin.Worker, pattern: Lapin.Pattern.HelloWorld
-end
-
-defmodule LapinTest.WorkQueueWorker do
-  use Lapin.Worker, pattern: Lapin.Pattern.WorkQueue
+defmodule LapinTest.HelloWorld do
+  use Lapin.Connection
 end
